hie-integration branch, squashed for rebase
- Expose a flag to force GHC to recompile the given targets

This is useful if the client app (e.g. HaRe or HIE) has a GHC hook set to pick
up info from the compilation.

- Improve ForceRecomp processing

1. Apply it after depanal
2. Undo the Opt_ForceRecomp flag after the load is complete

- Always allow updating the hook functions when loading a target

Regardless of whether it is a new session or not.

- Update the hooks in the modsummary for the selected targets

So that it actually gets used.

Which also means it is only activated for the specified targets, so the hook
function itself can be simpler, does not need to check whether it must do
special processing or not.

- Make sure the hook is called for interpreted mode loads

- Only load mapped files from the current component

- Better identification of changed DynFlags in a session

It needs an improvement of the output, in time.

- Don't load all mapped files

- Split into ghc-mod-core library and ghc-mod

First pass, each build, most tests pass for ghc-mod, missing 3, which I suspect
are from the file mapping changes made recently in the hie branch.

- Add GhcMod.ModuleLoader

This exposes getTypecheckedModuleGhc, for use in tooling based on ghc-mod-core

- Bring the module caching layer over into ModuleLoader

- Bring in balance of functions from HIE

- Move genLocMap etc here

- Remove TypeApplication for backward compatibility

- Ensure GHC 7.10.3 compatibility

- Clean up documentation

- Generalized getTypecheckedModuleGhc

- Added percent decoding for file URIs

- Added percent decoding for file URIs part 2

- Added percent decoding for file URIs part 3

- Export getTypecheckedModuleGhc'